Add files via upload
created an anonymous income comparison program.
The program must start by printing “Anonymous Income Comparison Program” to the screen.

It must then print “Person 1” to the screen and get the following details:

Hourly Rate

Hours worked per week

It must then print “Person 2” to the screen and then get the following details:

Hourly rate

Hours worked per week

It must then print to the screen “Annual salary of Person 1:” and write the exact salary below it.

It must then print to the screen “Annual salary of Person 2:” and write the exact salary below it.

It must then print to the screen “Person 1 makes more money than Person 2” and write the true or false value of this statement below it. 

Add comments to each line or block of your code to explain what it does exactly, so that another developer could read and understand your code.