Add files via upload
Console App Assignment Part 1
Create a console app the does/contains the following:

A one-dimensional array of strings.

Ask the user to input some text.

A loop that iterates through each string in the array and adds the user's text input to the end of each string. This step will not output anything to the console, but will update each array element by appending the user's text.

Then create a second loop that prints off each string in the array one at a time.

Console App Assignment Part 2
Add the following to your console app and perform these actions:

An infinite loop.

Save your code.

Fix the infinite loop so that it will execute properly.

Console App Assignment Part 3
Add the following to your console app and perform these actions:

A loop where the comparison that’s used to determine whether to continue iterating the loop is a “<” operator.

Save and execute your code.

Add another loop where the comparison that’s used to determine whether to continue iterating the loop is a “<=” operator.

Console App Assignment Part 4
Add the following to your console app and perform these actions:

A list of strings where each item in the list is unique.

Ask the user to input text to search for in the list.

A loop that iterates through the list and then displays the index of the list item that contains matching text on the screen.

Add code to check if the user put in text that isn't on the list and, if they did, tell the user their input is not on the list.

Add code that stops the loop from executing once a match has been found.Console App Assignment Part 4
Add the following to your console app and perform these actions:

A list of strings where each item in the list is unique.

Ask the user to input text to search for in the list.

A loop that iterates through the list and then displays the index of the list item that contains matching text on the screen.

Add code to check if the user put in text that isn't on the list and, if they did, tell the user their input is not on the list.

Add code that stops the loop from executing once a match has been found.


Console App Assignment Part 5
Add the following to your console app and perform these actions:

A list of strings that has at least two identical strings in the list. Ask the user to select text to search for in the list.

Create a loop that iterates through the list and then displays the indices of the list that contain matching text on the screen. Ensure to remove any break statements that may prevent your code from returning multiple matches.

Add code to the loop to check if the user put in text that isn't on the list and, if they did, tells the user their input is not on the list. (You do not need to provide any additional chances for the user to enter text.)


Console App Assignment Part 6

Add the following to your console app and perform these actions:

Create a list of strings that has at least two identical strings in the list. 

Create a foreach loop that evaluates each item in the list, and displays a message showing the string and whether or not it has already appeared in the list. For example, if you had a list of letters: {“A”, “B”, “C”, “D”, “C”}, you could have the following print to the screen:

A - this item is unique
B - this item is unique
C - this item is unique
D - this item is unique
C - this item is a duplicate